# add tracing functionality