--- a/export_test.go
+++ b/export_test.go
@@ -2,8 +2,9 @@ package macaroon
 
 var (
 	AddThirdPartyCaveatWithRand = (*Macaroon).addThirdPartyCaveatWithRand
-	MaxPacketV1Len              = maxPacketV1Len
 )
+
+type MacaroonJSONV2 macaroonJSONV2
 
 // SetVersion sets the version field of m to v;
 // usually so that we can compare it for deep equality with