This feature, mainly for debugging, allows a caller to
find out details of the macaroon verification. It
This is intended particularly for debugging divergent
macaroon implementations, but could also be useful
for other uncommon macaroon operations, such
as removing caveats given knowledge of a root key.

	benchmark                    old ns/op     new ns/op     delta
	BenchmarkNew-4               3125          3133          +0.26%
	BenchmarkAddCaveat-4         2831          2804          -0.95%
	BenchmarkVerifyLarge-4       60399         59982         -0.69%
	BenchmarkVerifySmall-4       3925          3914          -0.28%
	BenchmarkMarshalJSON-4       2039          2017          -1.08%
	BenchmarkUnmarshalJSON-4     3863          3843          -0.52%